add: adds links to today's goal in lobby
Links to 'Videos', 'Tasks', 'Slides' are conditionally added to the 'Today's Goal' section in Lobby. The computed property 'todaysGoal' was removed. Instead most of it's function is now part of the 'getScheduleDetailsToday' method, connecting schedule.vue with the according schedule-details.vue entry and creating an object 'today' with the details and links.